pyproject.toml
[project]
name = "grblwheel"
version = "0.1.0"
description = "GRBL jog wheel and G-code sender for Raspberry Pi and Windows"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pyserial>=3.5",
    "pyserial-asyncio>=0.6",
    "pyyaml>=6.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
gpio = ["pigpio>=1.78"]
dev = ["pytest>=7.0", "httpx>=0.25"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["backend"]
